refactor: reverted to int

+++ Pulse.API/Responses/ApiResponse.cs
@@ -14,7 +14,7 @@ public sealed class ApiResponse<T> : ApiResponse
 
 public sealed class ApiPagination
 {
-    public long Page { get; init; }
+    public int Page { get; init; }
     public int PageSize { get; init; }
     public int TotalCount { get; init; }
     public int TotalPages => PageSize <= 0 ? 0 : (int)Math.Ceiling((double)TotalCount / PageSize);
